## Added code to demonstrate the use of an if...else statement.
Added code to demonstrate the use of multiple nested
statements.

Added code to demonstrate the use of a logical or
operator.".

Added code to demonstrate the use of a logical and
operator.

Added code to demonstrate the use of a while loop with
ASCII art

Added code to demonstrate the use of a while loop with
len function."