## Added code to demonstrate the use of an if...else statement.
Added code to demonstrate the use of multiple nested
statements.

Added code to demonstrate the use of a logical or
operator.".

Added code to demonstrate the use of a logical and
operator.

Added code to demonstrate the use of a while loop with
ASCII art

"Added code to demonstrate counting down with a for loop.".
Added code to

demonstrate the use of a while loop with
len function."

"Added code to demonstrate the use of summing with a
while loop

"Added code to demonstrate the use of a while loop in
performing calculations with user input.".

Added code to demonstrate the use of a simple for loop

"Added code to demonstrate the use of for loops in string
indexing.".

"Added code to demonstrate how to reverse a string using a
for loop."

"Added code to demonstrate the use of membership
operators with a for loop

"Added code to demonstrate the use of one loop nested is
another."

Added code to demonstrate the use of nested loops and
decisions."